fix: 🎨 Addition of functions in utils.ts and imports in correction.ts
